Hide implementations of HideableValue
