+++ distribution/lib/Standard/Base/0.0.0-dev/src/Enso_Cloud/Enso_Secret.enso
@@ -336,16 +336,16 @@ Derived_Secret_Value.from (that : Enso_Secret) = Derived_Secret_Value.Secret_Val
    ---
 as_hideable_value : Text | Enso_Secret | Derived_Secret_Value -> HideableValue
 as_hideable_value (value : Text | Enso_Secret | Derived_Secret_Value) = case value of
-    text : Text -> HideableValue.PlainValue.new text
+    text : Text -> HideableValue.plain text
     secret : Enso_Secret ->
         if secret.is_credential then
             Error.throw (Illegal_Argument.Error "A credential for "+secret.credential_service_name+" cannot be used as a regular secret value. Credentials can only be used in their corresponding services.")
-        HideableValue.SecretValue.new secret.id
+        HideableValue.secret secret.id
     Derived_Secret_Value.Plain_Text text -> as_hideable_value text
     Derived_Secret_Value.Secret_Value secret -> as_hideable_value secret
-    Derived_Secret_Value.Concat left right -> HideableValue.ConcatValues.new (as_hideable_value left) (as_hideable_value right)
-    Derived_Secret_Value.Base_64_Encode inner -> HideableValue.Base64EncodeValue.new (as_hideable_value inner)
-    Derived_Secret_Value.Interpret_As_Private_Key secret -> InterpretAsPrivateKey.new (as_hideable_value secret)
+    Derived_Secret_Value.Concat left right -> HideableValue.concat (as_hideable_value left) (as_hideable_value right)
+    Derived_Secret_Value.Base_64_Encode inner -> HideableValue.base64 (as_hideable_value inner)
+    Derived_Secret_Value.Interpret_As_Private_Key secret -> HideableValue.privateKey (as_hideable_value secret)
 
 ## ---
    private: true
