+++ std-bits/base/src/main/java/org/enso/base/enso_cloud/SecretValueResolver.java
@@ -9,17 +9,17 @@ sealed class SecretValueResolver permits EnsoSecretHelper, ExternalLibrarySecret
    */
   protected static String resolveValue(HideableValue value) {
     return switch (value) {
-      case HideableValue.PlainValue plainValue -> plainValue.value();
-      case HideableValue.SecretValue secretValue ->
+      case HideableImpl.PlainValue plainValue -> plainValue.value();
+      case HideableImpl.SecretValue secretValue ->
           EnsoSecretReader.INSTANCE.readSecret(secretValue.secretId());
-      case HideableValue.ConcatValues concatValues -> {
+      case HideableImpl.ConcatValues concatValues -> {
         String left = resolveValue(concatValues.left());
         String right = resolveValue(concatValues.right());
         yield left + right;
       }
-      case HideableValue.Base64EncodeValue base64EncodeValue ->
-          HideableValue.Base64EncodeValue.encode(resolveValue(base64EncodeValue.value()));
-      case InterpretAsPrivateKey pk ->
+      case HideableImpl.Base64EncodeValue base64EncodeValue ->
+          HideableImpl.Base64EncodeValue.encode(resolveValue(base64EncodeValue.value()));
+      case HideableImpl.InterpretAsPrivateKey pk ->
           throw new IllegalStateException(
               "InterpretAsPrivateKey can only be used in JDBC connections. This state should never"
                   + " be reached.");
